challenge_6/ai_service/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
tabulate>=0.9.0
